nearly done with P 3

+++ CSC-1300/program_3/Statistics.h
@@ -1,10 +1,10 @@
 #if !defined STATFINDER
 #define STATFINDER
 
-int* readData (const char* filename, int& size);
-double findMean (int* numbers, int size);
-double findMedian (int* numbers, int size);
-int findMode (int* numbers, int size);
-void displayInformation (float mean, float median, int mode);
+long* readData (const char* filename, long& size);
+double findMean (long* numbers, long size);
+double findMedian (long* numbers, long size);
+long findMode (long* numbers, long size);
+void displayInformation (double mean, double median, long mode);
 
 #endif
